pyproject.toml
[project]
name = "xlsx-converter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "python-docx>=1.2.0",
    "reportlab>=4.4.10",
]
